@@ -11,28 +11,31 @@
     "@types/jest": "=23.3.8",
     "@types/react": "=16.8.3",
     "@types/react-native": "=0.57.42",
-    "@types/react-native-restart": "0.0.0",
+    "@types/react-native-restart": "=0.0.0",
+    "@types/react-native-sortable-list": "=0.0.7",
     "@types/react-native-vector-icons": "=4.6.4",
     "@types/react-navigation": "=2.13.7",
     "@types/react-redux": "=6.0.0",
     "@types/react-swipeable-views": "=0.12.2",
     "@types/react-test-renderer": "=16.0.3",
-    "babel-jest": "=24.5.0",
+    "babel-jest": "=24.6.0",
+    "detox": "=12.1.3",
     "eslint": "=5.15.3",
     "form-data": "=2.3.3",
-    "jest": "=24.5.0",
+    "jest": "=24.6.0",
     "jest-fetch-mock": "=1.6.6",
     "jest-serializer": "=23.0.1",
     "metro-react-native-babel-preset": "0.53.1",
-    "mocha": "=5.2.0",
+    "mocha": "=6.1.2",
     "mock-async-storage": "=2.0.3",
     "node-fetch": "=2.3.0",
     "react-native-asset": "=1.1.2",
     "react-native-debugger-open": "=0.3.17",
     "react-native-svg-transformer": "=0.12.1",
     "react-native-typescript-transformer": "=1.2.10",
+    "react-native-version": "=3.1.0",
     "react-test-renderer": "16.8.3",
-    "typescript": "=3.0.3"
+    "typescript": "=3.5.1"
   },
   "scripts": {
     "start": "react-native start",
