@@ -1,8 +1,18 @@
 const detox = require('detox');
 const config = require('../package.json').detox;
+const adapter = require('detox/runners/mocha/adapter');
 
 before(async () => {
-  await detox.init(config);
+  await detox.init(config, {launchApp: false});
+  await device.launchApp({permissions: {notifications: 'YES'}});
+});
+
+beforeEach(async function () {
+  await adapter.beforeEach(this);
+});
+
+afterEach(async function () {
+  await adapter.afterEach(this);
 });
 
 after(async () => {
