@@ -1,8 +1,14 @@
+import * as util from 'util';
+import { Utils } from './Utils';
+
 type Logger = (s: string) => void;
 
 export class Debug {
-    public static setDebug(isDebug: boolean) {
-        Debug.isDebug = isDebug;
+    public static isDebugMode = false;
+    public static useColors = true;
+
+    public static setDebugMode(isDebug: boolean) {
+        Debug.isDebugMode = isDebug;
     }
 
     public static setMaxLength(length: number) {
@@ -10,9 +16,20 @@ export class Debug {
     }
 
     public static log(...args: any[]) {
-        if (__DEV__ && Debug.isDebug) {
-            // tslint:disable-next-line:no-console
-            console.log.call(console, ...args);
+        if (__DEV__ && Debug.isDebugMode) {
+            if (Utils.isNodeJS() &&
+                args.length === 2 &&
+                typeof args[0] === 'string' &&
+                typeof args[1] === 'object'
+            ) {
+                const name = args[0];
+                const obj = args[1];
+                // tslint:disable-next-line:no-console
+                console.log(name, util.inspect(obj, false, null, Debug.useColors));
+            } else {
+                // tslint:disable-next-line:no-console
+                console.log.call(console, ...args);
+            }
         }
         const maxLengthArgs = args.map((value) => {
             const stringValue = (value instanceof Error)
@@ -30,12 +47,11 @@ export class Debug {
         Debug.loggers.push(logger);
     }
 
-    private static isDebug = false;
     private static maxLength = 1000;
     private static loggers: Logger[] = [];
 
     private static fullLog(...args: any[]) {
-        if (Debug.isDebug) {
+        if (Debug.isDebugMode) {
             const logLine = args.map(arg => '' + arg).join(' ');
             for (const logger of Debug.loggers) {
                 logger(logLine);
