@@ -2,7 +2,7 @@ import { ActionsUnion } from './types';
 import { createAction } from './actionHelpers';
 import { Feed } from '../models/Feed';
 import { ContentFilter } from '../models/ContentFilter';
-import { getAppStateFromSerialized, migrateAppStateToCurrentVersion } from '../reducers';
+import { migrateAppStateToCurrentVersion } from '../reducers';
 import { AppState } from '../reducers/AppState';
 import { RSSPostManager } from '../RSSPostManager';
 import { Post, PublicPost } from '../models/Post';
@@ -19,13 +19,15 @@ import {
 } from '../social/api';
 import * as Swarm from '../swarm/Swarm';
 import { PrivateIdentity } from '../models/Identity';
-import { restoreBackupToString } from '../BackupRestore';
 // @ts-ignore
 import { generateSecureRandom } from 'react-native-securerandom';
 import { isPostFeedUrl, loadRecentPosts, makeSwarmStorage, makeSwarmStorageSyncer, SwarmHelpers } from '../swarm-social/swarmStorage';
 import { resizeImageIfNeeded, resizeImageForPlaceholder } from '../ImageUtils';
 import { ReactNativeModelHelper } from '../models/ReactNativeModelHelper';
-import { FELFELE_ASSISTANT_URL } from '../reducers/defaultData';
+import { FELFELE_ASSISTANT_URL, FELFELE_FOUNDATION_URL } from '../reducers/defaultData';
+import { registerBackgroundTask } from '../helpers/backgroundTask';
+import { localNotification } from '../helpers/notifications';
+import { mergeUpdatedPosts } from '../helpers/postHelpers';
 
 export enum ActionTypes {
     ADD_CONTENT_FILTER = 'ADD-CONTENT-FILTER',
@@ -59,6 +61,7 @@ export enum ActionTypes {
     CHANGE_SETTING_SHOW_SQUARE_IMAGES = 'CHANGE-SETTING-SHOW-SQUARE-IMAGES',
     CHANGE_SETTING_SHOW_DEBUG_MENU = 'CHANGE-SETTING-SHOW-DEBUG-MENU',
     CHANGE_SETTING_SWARM_GATEWAY_ADDRESS = 'CHANGE-SETTING-SWARM-GATEWAY-ADDRESS',
+    CLEAN_FEEDS_FROM_OWN_FEEDS = 'CLEAN-FEEDS-FROM-OWN-FEEDS',
 }
 
 const InternalActions = {
@@ -66,6 +69,8 @@ const InternalActions = {
         createAction(ActionTypes.ADD_POST, { post }),
     increaseHighestSeenPostId: () =>
         createAction(ActionTypes.INCREASE_HIGHEST_SEEN_POST_ID),
+    addFeed: (feed: Feed) =>
+        createAction(ActionTypes.ADD_FEED, { feed }),
     addOwnFeed: (feed: LocalFeed) =>
         createAction(ActionTypes.ADD_OWN_FEED, { feed }),
     updateAuthorIdentity: (privateIdentity: PrivateIdentity) =>
@@ -74,6 +79,10 @@ const InternalActions = {
         createAction(ActionTypes.UPDATE_FEED_FAVICON, {feed, favicon}),
     appStateSet: (appState: AppState) =>
         createAction(ActionTypes.APP_STATE_SET, { appState }),
+    updateAuthorName: (name: string) =>
+        createAction(ActionTypes.UPDATE_AUTHOR_NAME, { name }),
+    updateAuthorImage: (image: ImageData) =>
+        createAction(ActionTypes.UPDATE_AUTHOR_IMAGE, { image }),
 };
 
 export const Actions = {
@@ -81,8 +90,6 @@ export const Actions = {
         createAction(ActionTypes.ADD_CONTENT_FILTER, { text, createdAt, validUntil }),
     removeContentFilter: (filter: ContentFilter) =>
         createAction(ActionTypes.REMOVE_CONTENT_FILTER, { filter }),
-    addFeed: (feed: Feed) =>
-        createAction(ActionTypes.ADD_FEED, { feed }),
     removeFeed: (feed: Feed) =>
         createAction(ActionTypes.REMOVE_FEED, { feed }),
     followFeed: (feed: Feed) =>
@@ -107,10 +114,6 @@ export const Actions = {
         createAction(ActionTypes.ADD_DRAFT, { draft }),
     removeDraft: () =>
         createAction(ActionTypes.REMOVE_DRAFT),
-    updateAuthorName: (name: string) =>
-        createAction(ActionTypes.UPDATE_AUTHOR_NAME, { name }),
-    updateAuthorImage: (image: ImageData) =>
-        createAction(ActionTypes.UPDATE_AUTHOR_IMAGE, { image }),
     appStateReset: () =>
         createAction(ActionTypes.APP_STATE_RESET),
     changeSettingSaveToCameraRoll: (value: boolean) =>
@@ -123,9 +126,19 @@ export const Actions = {
         createAction(ActionTypes.CHANGE_SETTING_SWARM_GATEWAY_ADDRESS, { value }),
     updateOwnFeed: (partialFeed: Partial<LocalFeed>) =>
         createAction(ActionTypes.UPDATE_OWN_FEED, { partialFeed }),
+    cleanFeedsFromOwnFeeds: (feedUrls: string[]) =>
+        createAction(ActionTypes.CLEAN_FEEDS_FROM_OWN_FEEDS, { feedUrls }),
 };
 
 export const AsyncActions = {
+    addFeed: (feed: Feed): Thunk => {
+        return async (dispatch, getState) => {
+            const ownFeeds = getState().ownFeeds.map(ownFeed => ownFeed.feedUrl);
+            if (!ownFeeds.includes(feed.feedUrl)) {
+                dispatch(InternalActions.addFeed(feed));
+            }
+        };
+    },
     cleanupContentFilters: (currentTimestamp: number = Date.now()): Thunk => {
         return async (dispatch, getState) => {
             const expiredFilters = getState().contentFilters.filter(filter =>
@@ -162,19 +175,7 @@ export const AsyncActions = {
             // TODO this is a hack, because we don't need a feed address
             const swarm = Swarm.makeReadableApi({user: '', topic: ''}, getState().settings.swarmGatewayAddress);
             const downloadedPosts = await loadPostsFromFeeds(swarm, feeds);
-            const uniqueAuthors = new Map<string, Author>();
-            downloadedPosts.map(post => {
-                if (post.author != null) {
-                    if (!uniqueAuthors.has(post.author.uri)) {
-                        uniqueAuthors.set(post.author.uri, post.author);
-                    }
-                }
-            });
-            const notUpdatedPosts = previousPosts.filter(post => post.author != null && !uniqueAuthors.has(post.author.uri));
-            const allPosts = notUpdatedPosts.concat(downloadedPosts);
-            const sortedPosts = allPosts.sort((a, b) => b.createdAt - a.createdAt);
-            const posts = sortedPosts.map((post, index) => ({...post, _id: index}));
-
+            const posts = mergeUpdatedPosts(downloadedPosts, previousPosts);
             dispatch(Actions.updateRssPosts(posts));
         };
     },
@@ -207,7 +208,7 @@ export const AsyncActions = {
                     ...localFeed,
                     postCommandLog: updatedPostCommandLog,
                 }));
-                dispatch(AsyncActions.syncPostCommandLogs(localFeed));
+                dispatch(AsyncActions.syncLocalFeed(localFeed));
             }
             dispatch(Actions.deletePost(post));
         };
@@ -282,6 +283,7 @@ export const AsyncActions = {
                 autoShare: true,
             };
             dispatch(InternalActions.addOwnFeed(ownFeed));
+            dispatch(AsyncActions.syncLocalFeed(ownFeed));
         };
     },
     shareOwnPost: (post: Post): Thunk => {
@@ -297,10 +299,10 @@ export const AsyncActions = {
                 ...localFeed,
                 postCommandLog: updatedPostCommandLog,
             }));
-            dispatch(AsyncActions.syncPostCommandLogs(localFeed));
+            dispatch(AsyncActions.syncLocalFeed(localFeed));
         };
     },
-    syncPostCommandLogs: (feed: LocalFeed): Thunk => {
+    syncLocalFeed: (feed: LocalFeed): Thunk => {
         return async (dispatch, getState) => {
             Debug.log('syncPostCommandLogs', 'feed', feed);
             const localFeed = getState().ownFeeds.find(ownFeed => ownFeed.feedUrl === feed.feedUrl);
@@ -352,7 +354,7 @@ export const AsyncActions = {
                         ...getState().author.image,
                         uri: storageSyncUpdate.recentPostFeed.authorImage.uri,
                     };
-                    dispatch(Actions.updateAuthorImage(authorImage));
+                    dispatch(InternalActions.updateAuthorImage(authorImage));
                 }
 
                 // Re-check if there were an update to the command log during syncing
@@ -374,10 +376,12 @@ export const AsyncActions = {
                     isSyncing: false,
                 }));
 
-                if (getPreviousCommandEpochFromLog(mergedPostCommandLog) == null) {
+                if (mergedPostCommandLog.commands.length > 0 &&
+                    getPreviousCommandEpochFromLog(mergedPostCommandLog) == null
+                ) {
                     Debug.log('syncPostCommandLogs', 'waiting for resyncing');
                     await Utils.waitMillisec(60 * 1000);
-                    dispatch(AsyncActions.syncPostCommandLogs(localFeedAfterUpdate));
+                    dispatch(AsyncActions.syncLocalFeed(localFeedAfterUpdate));
                 }
             } catch (e) {
                 Debug.log('syncPostCommandLogs: ', 'error', e);
@@ -403,18 +407,78 @@ export const AsyncActions = {
             }
         };
     },
+    createUser: (name: string, image: ImageData): Thunk => {
+        return async (dispatch, getState) => {
+            await dispatch(AsyncActions.chainActions([
+                AsyncActions.updateProfileName(name),
+                AsyncActions.updateProfileImage(image),
+                AsyncActions.createUserIdentity(),
+                AsyncActions.createOwnFeed(),
+            ]));
+        };
+    },
     createUserIdentity: (): Thunk => {
         return async (dispatch, getState) => {
             const privateIdentity = await Swarm.generateSecureIdentity(generateSecureRandom);
             dispatch(InternalActions.updateAuthorIdentity(privateIdentity));
         };
     },
-    restoreFromBackup: (backupText: string, secretHex: string): Thunk => {
+    restoreAppStateFromBackup: (appState: AppState): Thunk => {
         return async (dispatch, getState) => {
-            const serializedAppState = await restoreBackupToString(backupText, secretHex);
-            const appState = await getAppStateFromSerialized(serializedAppState);
             const currentVersionAppState = await migrateAppStateToCurrentVersion(appState);
             dispatch(InternalActions.appStateSet(currentVersionAppState));
+        };
+    },
+    updateProfileName: (name: string): Thunk => {
+        return async (dispatch, getState) => {
+            dispatch(InternalActions.updateAuthorName(name));
+            if (getState().ownFeeds.length > 0) {
+                const ownFeed = getState().ownFeeds[0];
+                dispatch(Actions.updateOwnFeed({
+                    feedUrl: ownFeed.feedUrl,
+                    name,
+                }));
+                dispatch(AsyncActions.syncLocalFeed(ownFeed));
+            }
+        };
+    },
+    updateProfileImage: (image: ImageData): Thunk => {
+        return async (dispatch, getState) => {
+            dispatch(InternalActions.updateAuthorImage(image));
+            if (getState().ownFeeds.length > 0) {
+                const ownFeed = getState().ownFeeds[0];
+                dispatch(Actions.updateOwnFeed({
+                    feedUrl: ownFeed.feedUrl,
+                    authorImage: image,
+                    favicon: undefined,
+                }));
+                dispatch(AsyncActions.syncLocalFeed(ownFeed));
+            }
+        };
+    },
+    registerBackgroundTasks: (): Thunk => {
+        return async (dispatch, getState) => {
+            const foundationFeeds = getState().feeds.filter(feed => feed.feedUrl === FELFELE_FOUNDATION_URL);
+            if (foundationFeeds.length > 0) {
+                const foundationFeed = foundationFeeds[0];
+                const getLatestPostCreateTime = (posts: Post[]) => posts.length > 0
+                    ? posts[0].createdAt
+                    : 0
+                ;
+                const backgroundTaskIntervalMinutes = 12 * 60;
+                registerBackgroundTask(backgroundTaskIntervalMinutes, async () => {
+                    const previousPosts = getState().rssPosts.filter(post => post.author != null && post.author.uri ===  foundationFeed.feedUrl);
+                    const previousSortedPosts = previousPosts.sort((a, b) => b.createdAt - a.createdAt);
+                    const address = Swarm.makeFeedAddressFromBzzFeedUrl(foundationFeeds[0].feedUrl);
+                    const swarm = Swarm.makeReadableApi(address, getState().settings.swarmGatewayAddress);
+                    const recentPosts = await loadRecentPosts(swarm, foundationFeeds);
+                    if (getLatestPostCreateTime(recentPosts) > getLatestPostCreateTime(previousSortedPosts)) {
+                        const posts = mergeUpdatedPosts(recentPosts, previousPosts);
+                        dispatch(Actions.updateRssPosts(posts));
+                        localNotification('There is a new version available!');
+                    }
+                });
+            }
         };
     },
 };
