@@ -1,16 +1,32 @@
 import * as React from 'react';
-import { View, StyleSheet, Clipboard, Alert, ShareContent, ShareOptions, Share, SafeAreaView } from 'react-native';
+import {
+    View,
+    StyleSheet,
+    Clipboard,
+    ShareContent,
+    ShareOptions,
+    Share,
+} from 'react-native';
+import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
+
 import { NavigationHeader } from './NavigationHeader';
 import { SimpleTextInput } from './SimpleTextInput';
 import { Debug } from '../Debug';
-import { Colors, DefaultNavigationBarHeight } from '../styles';
-import { Button } from './Button';
-import { createBackupFromString } from '../BackupRestore';
+import { Colors, ComponentColors, DefaultNavigationBarHeight, defaultMediumFont } from '../styles';
+import {
+    backupToSwarm,
+    encryptBackupLinkData,
+    generateBackupRandomSecret,
+} from '../helpers/backup';
 import { DateUtils } from '../DateUtils';
 import { getSerializedAppState } from '../reducers';
 import { AppState } from '../reducers/AppState';
-import { stringToHex } from '../conversion';
 import { TypedNavigation } from '../helpers/navigation';
+import * as Swarm from '../swarm/Swarm';
+import { HexString } from '../helpers/opaqueTypes';
+import { FragmentSafeAreaViewWithoutTabBar } from '../ui/misc/FragmentSafeAreaView';
+import { TouchableView } from './TouchableView';
+import { MediumText } from '../ui/misc/text';
 
 export interface StateProps {
     navigation: TypedNavigation;
@@ -23,20 +39,54 @@ export interface DispatchProps {
 export type Props = StateProps & DispatchProps;
 
 export interface State {
-    backupText: string;
-    secretText: string;
+    backupPassword: string;
+    randomSecret: HexString;
+    contentHash: HexString;
+    backupData: HexString;
     serializedAppState?: string;
 }
 
+const BackupButton = (props: {
+    onPress: () => void;
+    enabled: boolean;
+}) => {
+    const onPress = props.enabled
+        ? props.onPress
+        : undefined
+    ;
+    const color = props.enabled
+        ? Colors.BRAND_PURPLE
+        : Colors.LIGHT_GRAY
+    ;
+    return (
+        <TouchableView onPress={onPress} style={styles.buttonContainer}>
+            <View style={styles.buttonIcon}>
+                <Icon name='cloud-upload' color={color} size={24} />
+            </View>
+            <MediumText style={[styles.buttonLabel, {color}]}>Backup</MediumText>
+        </TouchableView>
+    );
+};
+
 export class Backup extends React.PureComponent<Props, State> {
-    public state = {
-        backupText: '',
-        secretText: '',
+    public state: State = {
+        backupPassword: '',
+        randomSecret: '' as HexString,
+        contentHash: '' as HexString,
+        backupData: '' as HexString,
         serializedAppState: undefined,
     };
 
+    public componentDidMount = async () => {
+        const randomSecret = await generateBackupRandomSecret();
+
+        this.setState({
+            randomSecret,
+        });
+    }
+
     public render = () => (
-        <SafeAreaView style={styles.mainContainer}>
+        <FragmentSafeAreaViewWithoutTabBar>
             <NavigationHeader
                 title='Backup'
                 navigation={this.props.navigation}
@@ -45,40 +95,46 @@ export class Backup extends React.PureComponent<Props, State> {
                 <SimpleTextInput
                     style={styles.secretTextInput}
                     multiline={true}
-                    numberOfLines={4}
+                    numberOfLines={1}
                     placeholder='Enter your backup password here'
                     autoCapitalize='none'
                     autoCorrect={false}
-                    onChangeText={async (text) => await this.setSecretText(text)}
+                    defaultValue={this.state.backupPassword}
+                    onChangeText={async (text) => await this.setBackupPassword(text)}
                 />
-                <Button style={styles.backupButton} text='Backup' onPress={async () => await this.onBackupData()} />
             </View>
+            <BackupButton
+                onPress={async () => await this.onBackupData()}
+                enabled={this.state.contentHash === ''}
+            />
             <SimpleTextInput
                 style={styles.backupTextInput}
                 editable={false}
-                defaultValue={this.state.backupText}
-                placeholder='Loading backup...'
+                defaultValue={this.getBackupText()}
+                placeholder='Saving backup...'
                 multiline={true}
             />
-        </SafeAreaView>
+        </FragmentSafeAreaViewWithoutTabBar>
     )
 
-    private setSecretText = async (text: string) => {
-        const secretHex = stringToHex(this.state.secretText);
-        const serializedAppState = await this.getOrLoadSerializedAppState();
-        const backupText = await createBackupFromString(serializedAppState, secretHex);
-
-        this.setState({
-            secretText: text,
-            backupText,
+    private completeSetState = <K extends keyof State>(state: ((prevState: Readonly<State>, props: Readonly<Props>) => (Pick<State, K> | State | null)) | (Pick<State, K> | State | null)): Promise<void> => {
+        return new Promise((resolve, reject) => {
+            this.setState(state, () => resolve());
         });
     }
 
-    private showShareDialog = async (message: string) => {
+    private setBackupPassword = async (text: string) => {
+        await this.completeSetState({
+            backupPassword: text,
+        });
+    }
+
+    private showShareDialog = async (backupData: HexString) => {
+        Debug.log('Backup.showShareDialog', 'backupLink', backupData);
         const title = 'Felfele backup ' + DateUtils.timestampToDateString(Date.now(), true);
         const content: ShareContent = {
             title,
-            message,
+            message: backupData,
         };
         const options: ShareOptions = {
         };
@@ -90,20 +146,33 @@ export class Backup extends React.PureComponent<Props, State> {
             return this.state.serializedAppState!;
         }
         const serializedAppState = await getSerializedAppState();
-        this.setState({
+        await this.completeSetState({
             serializedAppState,
         });
         return serializedAppState;
     }
 
+    private getBackupText = () => {
+        const backupText = `
+Random secret: ${this.state.randomSecret}
+Content hash: ${this.state.contentHash}
+Backup link: ${this.state.backupData}
+`;
+        return backupText;
+    }
+
     private onBackupData = async () => {
         try {
-            const secretHex = stringToHex(this.state.secretText);
+            const bzz = Swarm.makeBzzApi(this.props.appState.settings.swarmGatewayAddress);
             const serializedAppState = await this.getOrLoadSerializedAppState();
-            const emailBody = await createBackupFromString(serializedAppState, secretHex);
-            Debug.log('sendBackup body', emailBody);
-
-            this.showShareDialog(emailBody);
+            const contentHash = await backupToSwarm(bzz, serializedAppState, this.state.randomSecret);
+            const backupData = await encryptBackupLinkData(contentHash, this.state.randomSecret, this.state.backupPassword);
+            this.setState({
+                contentHash,
+                backupData,
+            });
+            Clipboard.setString(backupData);
+            await this.showShareDialog(backupData);
         } catch (e) {
             Debug.log('sendBackup error', e);
         }
@@ -114,32 +183,43 @@ const styles = StyleSheet.create({
     mainContainer: {
         height: '100%',
         flexDirection: 'column',
+        backgroundColor: ComponentColors.HEADER_COLOR,
     },
     backupTextInput: {
         fontSize: 10,
         flex: 1,
         padding: 3,
-        margin: 10,
         color: Colors.GRAY,
         backgroundColor: Colors.WHITE,
         marginBottom: DefaultNavigationBarHeight + 10,
-    },
-    backupButton: {
-        paddingVertical: 20,
-        paddingLeft: 20,
-        width: 100,
     },
     secretContainer: {
         flexDirection: 'row',
     },
     secretTextInput: {
-        fontSize: 12,
-        flex: 1,
-        padding: 3,
-        margin: 10,
-        borderRadius: 2,
-        height: 60,
+        width: '100%',
+        backgroundColor: 'white',
+        paddingHorizontal: 10,
+        paddingVertical: 14,
         color: Colors.DARK_GRAY,
+        fontSize: 14,
+        fontFamily: defaultMediumFont,
+        marginTop: 10,
+    },
+    buttonContainer: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        justifyContent: 'center',
         backgroundColor: Colors.WHITE,
+        margin: 10,
+        height: 44,
+    },
+    buttonIcon: {
+        alignItems: 'center',
+        paddingRight: 6,
+    },
+    buttonLabel: {
+        fontSize: 12,
+        color: Colors.BRAND_PURPLE,
     },
 });
