@@ -2,11 +2,13 @@ import * as React from 'react';
 import {
     View,
     StyleSheet,
-    SafeAreaView,
 } from 'react-native';
+import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 import { NavigationHeader } from './NavigationHeader';
-import { Button } from './Button';
-import { TypedNavigation, Routes } from '../helpers/navigation';
+import { TypedNavigation } from '../helpers/navigation';
+import { FragmentSafeAreaViewForTabBar } from '../ui/misc/FragmentSafeAreaView';
+import { TwoButton } from '../ui/buttons/TwoButton';
+import { Colors } from '../styles';
 
 export interface StateProps {
     navigation: TypedNavigation;
@@ -21,23 +23,29 @@ export interface State {
 }
 
 export const BackupRestore = (props: Props) => (
-    <SafeAreaView style={styles.mainContainer}>
+    <FragmentSafeAreaViewForTabBar>
         <NavigationHeader
             title='Backup & Restore'
             navigation={props.navigation}
         />
         <View style={styles.buttonContainer}>
-            <Button text='Backup' onPress={() => props.navigation.navigate('Backup', {})} />
-            <Button text='Restore' onPress={() => props.navigation.navigate('Restore', {})} />
+            <TwoButton
+                leftButton={{
+                    label: 'Backup',
+                    icon: <Icon name='cloud-upload' color={Colors.BRAND_PURPLE} size={24} />,
+                    onPress: () => props.navigation.navigate('Backup', {}),
+                }}
+                rightButton={{
+                    label: 'Restore',
+                    icon: <Icon name='cloud-download' color={Colors.BRAND_PURPLE} size={24} />,
+                    onPress: () => props.navigation.navigate('Restore', {}),
+                }}
+            />
         </View>
-    </SafeAreaView>
+    </FragmentSafeAreaViewForTabBar>
 );
 
 const styles = StyleSheet.create({
-    mainContainer: {
-        height: '100%',
-        flexDirection: 'column',
-    },
     buttonContainer: {
         flex: 1,
         flexDirection: 'column',
