@@ -1,19 +1,30 @@
 import * as React from 'react';
 import { NavigationHeader } from './NavigationHeader';
-import { Colors } from '../styles';
+import { Colors, ComponentColors } from '../styles';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import { View, StyleSheet, Linking, SafeAreaView } from 'react-native';
+import {
+    View,
+    StyleSheet,
+    ActivityIndicator,
+    Platform,
+    Text,
+    ScrollView,
+    KeyboardAvoidingView,
+} from 'react-native';
 import DeviceInfo from 'react-native-device-info';
-import { Button } from './Button';
 import { restartApp } from '../helpers/restart';
 import { BoldText, RegularText } from '../ui/misc/text';
 import { filteredLog, LogItem } from '../log';
 import { Version } from '../Version';
 
-// @ts-ignore
-import BugIcon from '../../images/bug.svg';
+import { Debug } from '../Debug';
+import { TypedNavigation } from '../helpers/navigation';
+import { SimpleTextInput } from './SimpleTextInput';
+import { WideButton } from '../ui/buttons/WideButton';
+import { FragmentSafeAreaViewWithoutTabBar, FragmentSafeAreaViewForTabBar } from '../ui/misc/FragmentSafeAreaView';
+import { TabBarPlaceholder } from '../ui/misc/TabBarPlaceholder';
+import { TwoButton } from '../ui/buttons/TwoButton';
 
-const BUG_REPORT_EMAIL_ADDRESS = 'bugreport@felfele.com';
 // personally identifiable information
 export const PIIKeys = [ 'privateKey', 'publicKey', 'address', 'name', 'localPath', 'user' ];
 
@@ -56,76 +67,241 @@ const piiFilteredLog = () => {
         ;
 };
 
-const getBugReportBody = (): string => {
-    const bugReportBody = `Please describe the bug: \n\n\n${deviceInfo()}Logs:\n${piiFilteredLog()}`;
-    return bugReportBody;
-};
+interface Props {
+    navigation?: TypedNavigation;
+    errorView: boolean;
+}
 
-const onPressSend = () =>
-    Linking.openURL(`mailto:${BUG_REPORT_EMAIL_ADDRESS}?subject=bugReport&body=${getBugReportBody()}`);
+interface State {
+    isSending: boolean;
+    feedbackText: string;
+}
 
-export const BugReportView = (props: { navigation?: any, errorView: boolean }) => {
-    return (
-        <SafeAreaView style={styles.mainContainer}>
-            <NavigationHeader
-                navigation={props.navigation}
-                title='Bug Report'
-                rightButton1={{
-                    onPress: onPressSend,
-                    label: <Icon
-                        name={'send'}
-                        size={20}
-                        color={Colors.BRAND_PURPLE}
-                    />,
-                }}
-            />
-            <View style={styles.contentContainer}>
-                <View style={styles.iconContainer}>
-                    <BugIcon
-                        width={29}
-                        height={29}
-                        fill={Colors.BRAND_PURPLE}
+export const BugReportViewWithTabBar = (props: Props) => (
+    <FragmentSafeAreaViewForTabBar>
+        <BugReportView {...props}/>
+    </FragmentSafeAreaViewForTabBar>
+);
+
+export const BugReportViewWithoutTabBar = (props: Props) => (
+    <FragmentSafeAreaViewWithoutTabBar>
+        <BugReportView {...props}/>
+    </FragmentSafeAreaViewWithoutTabBar>
+);
+
+class BugReportView extends React.Component<Props, State> {
+    public state: State = {
+        isSending: false,
+        feedbackText: '',
+    };
+
+    public render() {
+        return (
+            <KeyboardAvoidingView style={styles.keyboardAvoidingContainer}>
+                <NavigationHeader
+                    navigation={this.props.navigation}
+                    title='Bug Report'
+                />
+                <ScrollView contentContainerStyle={styles.contentContainer}>
+                    <View style={styles.iconContainer}>
+                        <Icon
+                            name={'bug'}
+                            size={36}
+                            color={Colors.BLACK}
+                        />
+                    </View>
+                    {this.props.errorView
+                        ? <BoldText style={[styles.text, { fontSize: 18 }]}>
+                            An error has occured!{'\n'}
+                            We need to restart the app.
+                        </BoldText>
+                        :  <BoldText style={[styles.text, { fontSize: 18 }]}>
+                            Yikes!{'\n'}
+                            What happened?
+                        </BoldText>
+                    }
+                    <RegularText style={[styles.text, { fontSize: 14 }]} textBreakStrategy='simple'>
+                        As we never collect information automatically, it would be truly helpful if you could take a moment to let us know what happened.
+                    </RegularText>
+                    <SimpleTextInput
+                        style={styles.textInput}
+                        multiline={true}
+                        numberOfLines={4}
+                        onChangeText={this.onChangeText}
+                        placeholder='Let us know what happened...'
+                        placeholderTextColor='gray'
+                        underlineColorAndroid='transparent'
                     />
-                </View>
-                {props.errorView &&
-                    <BoldText style={[styles.text, { fontSize: 18 }]}>
-                        Yikes!{'\n\n'}
-                        We are sorry, an error has occurred.{'\n'}
-                    </BoldText>
-                }
-                <RegularText style={[styles.text, { fontSize: 14 }]}>
-                    As we never collect information automatically, it would be truly helpful if you could take a moment to let us know what happened.
-                </RegularText>
-                <RegularText style={[styles.text, { fontSize: 14, color: Colors.BRAND_PURPLE }]}>
-                    By sending a bug report, you will share some of your information with us. You can review everything in your email client before sending.{'\n\n'}
-                    Tap on the Send button to continue.
-                </RegularText>
-                {props.errorView &&
-                    <Button style={styles.restartButton} text='Restart' onPress={restartApp} />
-                }
-            </View>
-        </SafeAreaView>
-    );
-};
+                    {!this.props.errorView
+                        ? <this.SendBugReportButton/>
+                        : <TwoButton
+                            leftButton={{
+                                icon:
+                                    <Icon
+                                        name={'refresh'}
+                                        size={24}
+                                        color={Colors.BRAND_PURPLE}
+                                    />
+                                ,
+                                label: 'RESTART',
+                                onPress: restartApp,
+                            }}
+                            rightButton={{
+                                icon: !this.state.isSending ?
+                                    <Icon
+                                        name={'send'}
+                                        size={24}
+                                        color={Colors.BRAND_PURPLE}
+                                    /> :
+                                    <ActivityIndicator size='small' color='grey' />
+                                ,
+                                label: 'SEND BUG REPORT',
+                                onPress: this.onPressSend,
+                            }}
+                        />
+                    }
+                    <RegularText style={[styles.text, { fontSize: 14, color: Colors.BRAND_PURPLE }]}>
+                        By sending a bug report, you will share some information (shown below) with us.
+                    </RegularText>
+                    <View style={styles.logContainer}>
+                        <Text style={styles.logText}>{this.getDeviceInfoAndLogs()}</Text>
+                    </View>
+                    <this.SendBugReportButton/>
+                    <TabBarPlaceholder/>
+                </ScrollView>
+            </KeyboardAvoidingView>
+        );
+    }
+
+    private SendBugReportButton = () => (
+        <WideButton
+            style={{marginBottom: 20}}
+            icon={!this.state.isSending ?
+                <Icon
+                    name={'send'}
+                    size={24}
+                    color={Colors.BRAND_PURPLE}
+                /> :
+                <ActivityIndicator size='small' color='grey' />
+            }
+            onPress={this.onPressSend}
+            label={'SEND BUG REPORT'}
+        />
+    )
+
+    private onChangeText = (feedbackText: string) => {
+        this.setState({ feedbackText });
+    }
+
+    private onPressSend = async () => {
+        this.setState({
+            isSending: true,
+        });
+
+        await this.sendBugReport();
+
+        this.setState({
+            isSending: false,
+            feedbackText: '',
+        });
+
+        if (this.props.navigation != null) {
+            this.props.navigation.goBack();
+        } else if (this.props.errorView) {
+            restartApp();
+        }
+    }
+
+    private getBugReportBody = (): string => {
+        return `User Feedback:
+
+${this.state.feedbackText}
+${this.getDeviceInfoAndLogs()}
+`;
+    }
+
+    private getDeviceInfoAndLogs = (): string => {
+        const bugReportBody = `Device Info:
+
+${deviceInfo()}
+Logs:
+
+${piiFilteredLog()}`;
+        return bugReportBody;
+    }
+
+    private sendBugReport = async () => {
+        try {
+            const response = await fetch('https://app.felfele.com/api/v1/bugreport/', {
+                headers: {
+                    'Content-Type': 'text/plain',
+                },
+                method: 'POST',
+                body: this.getBugReportBody(),
+            });
+            Debug.log('success sending bugreport', response.status);
+        } catch (e) {
+            Debug.log('error sending bugreport', e);
+        }
+    }
+}
+
+const fontFamily = Platform.OS === 'ios' ? 'Courier' : 'monospace';
 
 const styles = StyleSheet.create({
     mainContainer: {
-        backgroundColor: Colors.WHITE,
+        backgroundColor: ComponentColors.HEADER_COLOR,
+        flex: 1,
+    },
+    keyboardAvoidingContainer: {
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
         flex: 1,
     },
     contentContainer: {
-        paddingTop: 25,
-        flexDirection: 'column',
         alignItems: 'center',
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
     },
     iconContainer: {
+        paddingTop: 26,
+        paddingBottom: 30,
     },
     text: {
         textAlign: 'center',
-        maxWidth: '80%',
-        paddingTop: 50,
+        paddingHorizontal: 10,
+        paddingBottom: 10,
+    },
+    label: {
+        alignSelf: 'flex-start',
+        fontSize: 12,
+        paddingHorizontal: 10,
+        paddingTop: 9,
+        paddingBottom: 7,
+        color: Colors.GRAY,
+    },
+    logContainer: {
+        width: '100%',
+        marginVertical: 20,
+        backgroundColor: Colors.LIGHTER_GRAY,
+        paddingHorizontal: 10,
+        paddingVertical: 12,
+    },
+    logText: {
+        fontFamily: fontFamily,
+        fontSize: 14,
+        color: Colors.DARK_GRAY,
+        backgroundColor: Colors.LIGHTER_GRAY,
     },
     restartButton: {
         paddingTop: 50,
+    },
+    textInput: {
+        marginTop: 20,
+        marginBottom: 10,
+        padding: 10,
+        backgroundColor: Colors.WHITE,
+        fontSize: 18,
+        height: 190,
+        width: '100%',
+        textAlignVertical: 'top',
     },
 });
