@@ -7,27 +7,31 @@ import {
     ActivityIndicator,
     Dimensions,
     Clipboard,
-    SafeAreaView,
 } from 'react-native';
 import QRCode from 'react-native-qrcode-svg';
-import QRCodeScanner, { Event as ScanEvent } from 'react-native-qrcode-scanner';
+import QRCodeScanner from 'react-native-qrcode-scanner';
 
 import { RSSFeedManager } from '../RSSPostManager';
 import * as urlUtils from '../helpers/urlUtils';
 import { Feed } from '../models/Feed';
 import { SimpleTextInput } from './SimpleTextInput';
 import { Debug } from '../Debug';
-import { Colors } from '../styles';
+import { ComponentColors, Colors, defaultMediumFont } from '../styles';
 import * as Swarm from '../swarm/Swarm';
 import { downloadRecentPostFeed } from '../swarm-social/swarmStorage';
 import { NavigationHeader } from './NavigationHeader';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 import { unfollowFeed } from './FeedView';
-import { TypedNavigation, Routes } from '../helpers/navigation';
+import { TypedNavigation } from '../helpers/navigation';
+import { FragmentSafeAreaViewWithoutTabBar } from '../ui/misc/FragmentSafeAreaView';
+import { WideButton } from '../ui/buttons/WideButton';
+import { RegularText } from '../ui/misc/text';
+import { showShareFeedDialog } from '../helpers/shareDialogs';
+import { getFeedUrlFromFollowLink } from '../helpers/deepLinking';
 
-const QRCodeWidth = Dimensions.get('window').width * 0.6;
+const QRCodeWidth = Dimensions.get('window').width * 0.8;
 const QRCodeHeight = QRCodeWidth;
-const QRCameraWidth = Dimensions.get('window').width * 0.6;
+const QRCameraWidth = Dimensions.get('window').width;
 const QRCameraHeight = QRCameraWidth;
 
 interface FeedInfoState {
@@ -76,7 +80,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
         this.props.onAddFeed(feed);
     }
 
-    public async fetchFeed(onSuccess?: () => void, feedUrl?: string) {
+    public async fetchFeed(feedUrl?: string) {
         Debug.log('fetchFeed', 'this.state', this.state);
         if (this.state.loading === true) {
             return;
@@ -84,18 +88,15 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
 
         this.setState({
             loading: true,
-            activityText: 'Loading feed',
+            activityText: 'Loading channel...',
         });
 
-        const url = feedUrl != null ? feedUrl : this.state.url;
+        const url = this.tryGetFeedUrlFromFollowLink(feedUrl != null ? feedUrl : this.state.url);
         const feed = await this.fetchFeedFromUrl(url);
         if (feed != null && feed.feedUrl !== '') {
             this.setState({
                 loading: false,
             });
-            if (onSuccess != null) {
-                onSuccess();
-            }
             this.onAdd(feed);
             this.props.navigation.navigate('Feed', {
                 feedUrl: feed.feedUrl,
@@ -110,7 +111,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
         const isExistingFeed = this.props.feed.feedUrl.length > 0;
         const isFollowed = this.props.feed.followed;
 
-        const icon = (name: string) => <Icon name={name} size={20} color={Colors.DARK_GRAY} />;
+        const icon = (name: string, size: number = 20) => <Icon name={name} size={size} color={ComponentColors.NAVIGATION_BUTTON_COLOR} />;
         const button = (iconName: string, onPress: () => void) => ({
             label: icon(iconName),
             onPress,
@@ -118,17 +119,21 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
 
         const rightButton1 = isExistingFeed
             ? isFollowed
-                ? button('link-variant-off', this.onUnfollowFeed)
+                ? button('link-off', this.onUnfollowFeed)
                 : this.props.isKnownFeed
                     ? button('delete', this.onDelete)
                     : undefined
-            : button('download', async () => await this.fetchFeed())
+            : undefined
         ;
 
         return (
-            <SafeAreaView style={styles.mainContainer}>
+            <FragmentSafeAreaViewWithoutTabBar>
                 <NavigationHeader
-                    title={isExistingFeed ? 'Feed Info' : 'Add Feed'}
+                    title={isExistingFeed ? this.props.feed.name : 'Add channel'}
+                    leftButton={{
+                        label: icon('close', 24),
+                        onPress: () => this.props.navigation.goBack(null),
+                    }}
                     rightButton1={rightButton1}
                     navigation={this.props.navigation}
                 />
@@ -137,9 +142,10 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
                         defaultValue={this.state.url}
                         style={styles.linkInput}
                         onChangeText={(text) => this.setState({ url: text })}
-                        placeholder='Link of the feed'
+                        placeholder='Scan QR code or paste link here'
+                        placeholderTextColor={Colors.MEDIUM_GRAY}
                         autoCapitalize='none'
-                        autoFocus={true}
+                        autoFocus={false}
                         autoCorrect={false}
                         editable={!isExistingFeed}
                         returnKeyType='done'
@@ -157,7 +163,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
                         : <this.NewItemView showQRCamera={this.state.showQRCamera} />
                     }
                 </View>
-            </SafeAreaView>
+            </FragmentSafeAreaViewWithoutTabBar>
         );
     }
 
@@ -174,7 +180,6 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
                             cameraProps={{ratio: '1:1'}}
                         />
                     </View>
-                    <Text style={styles.qrCameraText}>You can scan a QR code too</Text>
                 </View>
             );
         } else {
@@ -190,10 +195,17 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
                     <QRCode
                         value={qrCodeValue}
                         size={QRCodeWidth}
-                        color={Colors.DARK_GRAY}
-                        backgroundColor={Colors.BACKGROUND_COLOR}
+                        color={Colors.BLACK}
+                        backgroundColor={ComponentColors.BACKGROUND_COLOR}
                     />
                 </View>
+                <RegularText style={styles.qrCodeText}>Show this QR code or use the link</RegularText>
+                <WideButton
+                    label='SHARE'
+                    icon={<Icon name='share' size={24} color={Colors.BRAND_PURPLE} />}
+                    style={{marginTop: 20}}
+                    onPress={async () => showShareFeedDialog(this.props.feed)}
+                />
             </View>
         );
     }
@@ -207,10 +219,18 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
                 this.setState({
                     url: link,
                 });
-                const clearClipboard = () => Clipboard.setString('');
-                await this.fetchFeed(clearClipboard, link);
+                Clipboard.setString('');
+                await this.fetchFeed(link);
             }
         }
+    }
+
+    private tryGetFeedUrlFromFollowLink = (followLink: string): string => {
+        const feedUrlFromFollowLink = getFeedUrlFromFollowLink(followLink);
+        return feedUrlFromFollowLink != null
+            ? feedUrlFromFollowLink
+            : followLink
+        ;
     }
 
     private fetchFeedFromUrl = async (url: string): Promise<Feed | null> => {
@@ -224,7 +244,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
                 Debug.log('fetchFeedFromUrl', 'url', url);
                 const canonicalUrl = urlUtils.getCanonicalUrl(url);
                 Debug.log('fetchFeedFromUrl', 'canonicalUrl', canonicalUrl);
-                const feed = await this.fetchRSSFeedFromUrl(canonicalUrl);
+                const feed = await RSSFeedManager.fetchFeedFromUrl(url);
                 Debug.log('fetchFeedFromUrl', 'feed', feed);
                 return feed;
             }
@@ -232,12 +252,6 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
             Debug.log(e);
             return null;
         }
-    }
-
-    private fetchRSSFeedFromUrl = async (url: string): Promise<Feed | null> => {
-        const feed = await RSSFeedManager.fetchFeedFromUrl(url);
-        Debug.log('fetchFeedFromUrl: feed: ', feed);
-        return feed;
     }
 
     private onUnfollowFeed = () => {
@@ -250,7 +264,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
             { text: 'Cancel', onPress: () => Debug.log('Cancel Pressed'), style: 'cancel' },
         ];
 
-        Alert.alert('Are you sure you want to delete the feed?',
+        Alert.alert('Are you sure you want to delete the channel?',
             undefined,
             options,
             { cancelable: true },
@@ -262,7 +276,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
             { text: 'Cancel', onPress: () => Debug.log('Cancel Pressed'), style: 'cancel' },
         ];
 
-        Alert.alert('Failed to load feed!',
+        Alert.alert('Failed to load channel!',
             undefined,
             options,
             { cancelable: true },
@@ -280,7 +294,7 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
             this.setState({
                 url: feedUrl,
             });
-            await this.fetchFeed(undefined, feedUrl);
+            await this.fetchFeed(feedUrl);
         } catch (e) {
             Debug.log(e);
         }
@@ -288,46 +302,27 @@ export class FeedInfo extends React.Component<Props, FeedInfoState> {
 }
 
 const styles = StyleSheet.create({
-    mainContainer: {
-        backgroundColor: Colors.WHITE,
-        flex: 1,
-        flexDirection: 'column',
-    },
     container: {
-        backgroundColor: Colors.BACKGROUND_COLOR,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
         flex: 1,
         flexDirection: 'column',
-    },
-    titleInfo: {
-        fontSize: 14,
-        color: '#8e8e93',
     },
     linkInput: {
         width: '100%',
         backgroundColor: 'white',
-        borderBottomColor: 'lightgray',
-        borderBottomWidth: 1,
-        borderTopColor: 'lightgray',
-        paddingHorizontal: 8,
-        paddingVertical: 8,
-        color: 'gray',
-        fontSize: 16,
-    },
-    deleteButtonContainer: {
-        backgroundColor: 'white',
-        width: '100%',
-        position: 'absolute',
-        bottom: 0,
-        left: 0,
         paddingHorizontal: 10,
-        paddingVertical: 10,
+        paddingVertical: 14,
+        color: Colors.DARK_GRAY,
+        fontSize: 14,
+        fontFamily: defaultMediumFont,
+        marginTop: 10,
     },
     centerIcon: {
         width: '100%',
         justifyContent: 'center',
         flexDirection: 'column',
         height: 100,
-        backgroundColor: Colors.BACKGROUND_COLOR,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
         paddingTop: 50,
     },
     qrCodeContainer: {
@@ -336,6 +331,12 @@ const styles = StyleSheet.create({
         height: QRCodeHeight,
         padding: 0,
         alignSelf: 'center',
+    },
+    qrCodeText: {
+        fontSize: 14,
+        color: Colors.GRAY,
+        marginTop: 20,
+        marginLeft: 10,
     },
     qrCameraContainer: {
         width: QRCameraWidth,
