@@ -1,11 +1,9 @@
 import * as React from 'react';
-import { View, StyleSheet, SafeAreaView, TouchableWithoutFeedback } from 'react-native';
-import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
+import { StyleSheet, SafeAreaView } from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 
 import { Feed } from '../models/Feed';
-import { ImageData } from '../models/ImageData';
-import { Colors } from '../styles';
+import { Colors, ComponentColors } from '../styles';
 import { NavigationHeader } from './NavigationHeader';
 import { SuperGridSectionList } from 'react-native-super-grid';
 import { GridCard, getGridCardSize } from '../ui/misc/GridCard';
@@ -14,6 +12,9 @@ import { MediumText } from '../ui/misc/text';
 import { TabBarPlaceholder } from '../ui/misc/TabBarPlaceholder';
 import { defaultImages } from '../defaultImages';
 import { TypedNavigation } from '../helpers/navigation';
+import { FragmentSafeAreaViewWithoutTabBar } from '../ui/misc/FragmentSafeAreaView';
+import { TwoButton } from '../ui/buttons/TwoButton';
+import { getFeedImage } from '../helpers/feedHelpers';
 
 export interface DispatchProps {
     onPressFeed: (feed: Feed) => void;
@@ -31,6 +32,7 @@ export interface StateProps {
     gatewayAddress: string;
     title: string;
     showExplore: boolean;
+    headerComponent?: React.ComponentType<any> | React.ReactElement<any> | null;
 }
 
 export class FeedGrid extends React.PureComponent<DispatchProps & StateProps & { children?: React.ReactNode}> {
@@ -38,8 +40,10 @@ export class FeedGrid extends React.PureComponent<DispatchProps & StateProps & {
         const itemDimension = getGridCardSize();
         const modelHelper = new ReactNativeModelHelper(this.props.gatewayAddress);
         return (
-            <View style={{ backgroundColor: Colors.BACKGROUND_COLOR, flex: 1 }}>
+            <SafeAreaView style={{ backgroundColor: ComponentColors.BACKGROUND_COLOR, flex: 1 }}>
                 {this.props.children}
+                {
+                // @ts-ignore - SuperGridSectionList is passing props to internal SectionList, typings is missing
                 <SuperGridSectionList
                     style={{ flex: 1 }}
                     spacing={10}
@@ -47,18 +51,14 @@ export class FeedGrid extends React.PureComponent<DispatchProps & StateProps & {
                     itemDimension={itemDimension}
                     sections={this.props.sections}
                     renderItem={({ item }: any) => {
-                        const image: ImageData = item.authorImage != null
-                            ? item.authorImage
-                            : { uri: item.favicon }
-                            ;
-                        const imageUri = modelHelper.getImageUri(image);
+                        const image = getFeedImage(item);
                         return (
                             <GridCard
                                 title={item.name}
-                                imageUri={imageUri}
+                                image={image}
                                 onPress={() => this.props.onPressFeed(item)}
                                 size={itemDimension}
-                                defaultImage={defaultImages.userCircle}
+                                defaultImage={defaultImages.defaultUser}
                                 modelHelper={modelHelper}
                             />
                         );
@@ -67,45 +67,47 @@ export class FeedGrid extends React.PureComponent<DispatchProps & StateProps & {
                         <MediumText style={styles.sectionHeader}>{section.title}</MediumText>
                     )}
                     // @ts-ignore - SuperGridSectionList is passing props to internal SectionList, typings is missing
-                    ListFooterComponent={<TabBarPlaceholder color={Colors.BACKGROUND_COLOR}/>}
+                    ListFooterComponent={<TabBarPlaceholder color={ComponentColors.BACKGROUND_COLOR}/>}
+                    ListHeaderComponent={this.props.headerComponent}
                 />
-            </View>
+                }
+            </SafeAreaView>
         );
     }
 }
 
+const ActionButtons = (openExplore: () => void, openAddChannel: () => void) => (
+    <TwoButton
+        leftButton={{
+            label: 'Add channel',
+            icon: <Icon name='account-plus' size={24} color={Colors.BRAND_PURPLE} />,
+            onPress: openAddChannel,
+        }}
+        rightButton={{
+            label: 'Explore',
+            icon: <Icon name='compass' size={24} color={Colors.BRAND_PURPLE}/>,
+            onPress: openExplore,
+        }}
+    />
+);
+
 export class FeedListEditor extends React.PureComponent<DispatchProps & StateProps> {
     public render() {
         return (
-            <SafeAreaView style={{ backgroundColor: Colors.WHITE, flex: 1 }}>
-                <FeedGrid {...this.props}>
+            <FragmentSafeAreaViewWithoutTabBar>
+                <FeedGrid
+                    headerComponent={this.props.showExplore
+                        ? ActionButtons(this.props.openExplore, this.onAddFeed)
+                        : undefined
+                    }
+                    {...this.props}
+                >
                     <NavigationHeader
                         navigation={this.props.navigation}
-                        rightButton1={{
-                            onPress: this.onAddFeed,
-                            label: <MaterialIcon name='add-box' size={24} color={Colors.BUTTON_COLOR} />,
-                        }}
                         title={this.props.title}
                     />
-                    {this.props.showExplore &&
-                    <TouchableWithoutFeedback
-                        onPress={() => this.props.openExplore()}
-                    >
-                        <View style={{
-                            backgroundColor: Colors.WHITE,
-                            height: 44,
-                            flexDirection: 'row',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                        }}>
-                            <View style={{ paddingTop: 1, paddingRight: 4 }}>
-                                <Icon name='compass' size={20} color={Colors.DARK_GRAY}/>
-                            </View>
-                            <MediumText style={{ fontSize: 12, color: Colors.DARK_GRAY }}>EXPLORE PUBLIC FEEDS</MediumText>
-                        </View>
-                    </TouchableWithoutFeedback>}
                 </FeedGrid>
-            </SafeAreaView>
+            </FragmentSafeAreaViewWithoutTabBar>
         );
     }
 
@@ -126,7 +128,7 @@ const styles = StyleSheet.create({
         paddingTop: 20,
         paddingBottom: 7,
         color: Colors.DARK_GRAY,
-        backgroundColor: Colors.BACKGROUND_COLOR,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
         fontSize: 14,
     },
 });
