@@ -4,14 +4,9 @@ import {
     KeyboardAvoidingView,
     StyleSheet,
     View,
-    Text,
     Image,
     TouchableOpacity,
     Dimensions,
-    Share,
-    ShareContent,
-    ShareOptions,
-    Platform,
     ScrollView,
     SafeAreaView,
 } from 'react-native';
@@ -20,20 +15,24 @@ import MaterialCommunityIcon from 'react-native-vector-icons/MaterialCommunityIc
 import { SimpleTextInput } from './SimpleTextInput';
 import { Author } from '../models/Author';
 import { ImageData } from '../models/ImageData';
-import { AsyncImagePicker } from '../AsyncImagePicker';
-import { Colors } from '../styles';
-import { NavigationHeader } from './NavigationHeader';
 import { Feed } from '../models/Feed';
+import { AsyncImagePicker } from '../AsyncImagePicker';
+import { ComponentColors, Colors } from '../styles';
+import { NavigationHeader } from './NavigationHeader';
 import { Debug } from '../Debug';
 import { ReactNativeModelHelper } from '../models/ReactNativeModelHelper';
-import { RowItem } from '../ui/misc/RowButton';
+import { RowItem } from '../ui/buttons/RowButton';
 import { RegularText } from '../ui/misc/text';
 import { TabBarPlaceholder } from '../ui/misc/TabBarPlaceholder';
 import { defaultImages } from '../defaultImages';
 import { DEFAULT_AUTHOR_NAME } from '../reducers/defaultData';
 import { TypedNavigation } from '../helpers/navigation';
+import { LocalFeed } from '../social/api';
+import { showShareFeedDialog } from '../helpers/shareDialogs';
+import { TwoButton } from '../ui/buttons/TwoButton';
+import { ImageDataView } from '../components/ImageDataView';
 
-const defaultUserImage = defaultImages.userCircle;
+const defaultUserImage = defaultImages.defaultUser;
 
 export interface DispatchProps {
     onUpdateAuthor: (text: string) => void;
@@ -43,7 +42,7 @@ export interface DispatchProps {
 
 export interface StateProps {
     author: Author;
-    ownFeed?: Feed;
+    ownFeed?: LocalFeed;
     navigation: TypedNavigation;
     gatewayAddress: string;
 }
@@ -63,25 +62,9 @@ const generateQRCodeValue = (feed?: Feed): string => {
     return feed.url;
 };
 
-const showShareDialog = async (feed?: Feed) => {
-    const url = feed != null ? feed.url : '';
-    const title = 'Share your feed';
-    const message = Platform.OS === 'android' ? url : undefined;
-    const content: ShareContent = {
-        url,
-        title,
-        message,
-    };
-    const options: ShareOptions = {
-    };
-    await Share.share(content, options);
-};
-
 export const IdentitySettings = (props: DispatchProps & StateProps) => {
     const qrCodeValue = generateQRCodeValue(props.ownFeed);
     const modelHelper = new ReactNativeModelHelper(props.gatewayAddress);
-    const authorImageUri = modelHelper.getImageUri(props.author.image);
-    Debug.log('IdentitySettings', authorImageUri);
     return (
         <SafeAreaView style={styles.safeAreaContainer}>
             <KeyboardAvoidingView style={styles.mainContainer}>
@@ -92,9 +75,9 @@ export const IdentitySettings = (props: DispatchProps & StateProps) => {
                                 label: <MaterialCommunityIcon
                                     name={'share'}
                                     size={20}
-                                    color={Colors.DARK_GRAY}
+                                    color={ComponentColors.NAVIGATION_BUTTON_COLOR}
                                 />,
-                                onPress: async () => showShareDialog(props.ownFeed),
+                                onPress: async () => showShareFeedDialog(props.ownFeed),
                             }
                             : undefined
                     }
@@ -109,12 +92,11 @@ export const IdentitySettings = (props: DispatchProps & StateProps) => {
                         }}
                         style={styles.imagePickerContainer}
                     >
-                        <Image
-                            source={authorImageUri === ''
-                                ? defaultUserImage
-                                : { uri: authorImageUri }
-                            }
+                        <ImageDataView
+                            source={props.author.image}
+                            defaultImage={defaultUserImage}
                             style={styles.imagePicker}
+                            modelHelper={modelHelper}
                         />
                     </TouchableOpacity>
                     <RegularText style={styles.tooltip}>{NAME_LABEL}</RegularText>
@@ -144,11 +126,30 @@ export const IdentitySettings = (props: DispatchProps & StateProps) => {
                             <QRCode
                                 value={qrCodeValue}
                                 size={QRCodeWidth}
-                                color={Colors.DARK_GRAY}
-                                backgroundColor={Colors.BACKGROUND_COLOR}
+                                color={Colors.BLACK}
+                                backgroundColor={ComponentColors.BACKGROUND_COLOR}
                             />
                         </View>
                     }
+                    <TwoButton
+                        leftButton={{
+                            label: 'Share',
+                            icon: <MaterialCommunityIcon name='share' size={24} color={Colors.BRAND_PURPLE} />,
+                            onPress: async () => showShareFeedDialog(props.ownFeed),
+                        }}
+                        rightButton={{
+                            label: 'Add channel',
+                            icon: <MaterialCommunityIcon name='account-plus' size={24} color={Colors.BRAND_PURPLE} />,
+                            onPress: () => props.navigation.navigate('FeedInfo', {
+                                feed: {
+                                    name: '',
+                                    url: '',
+                                    feedUrl: '',
+                                    favicon: '',
+                                },
+                           }),
+                        }}
+                    />
                 </ScrollView>
                 <TabBarPlaceholder/>
             </KeyboardAvoidingView>
@@ -157,7 +158,7 @@ export const IdentitySettings = (props: DispatchProps & StateProps) => {
 };
 
 const openImagePicker = async (onUpdatePicture: (imageData: ImageData) => void) => {
-    const imageData = await AsyncImagePicker.launchImageLibrary();
+    const imageData = await AsyncImagePicker.showImagePicker();
     if (imageData != null) {
         onUpdatePicture(imageData);
     }
@@ -165,11 +166,11 @@ const openImagePicker = async (onUpdatePicture: (imageData: ImageData) => void) 
 
 const styles = StyleSheet.create({
     safeAreaContainer: {
-        backgroundColor: Colors.WHITE,
+        backgroundColor: ComponentColors.HEADER_COLOR,
         flex: 1,
     },
     mainContainer: {
-        backgroundColor: Colors.BACKGROUND_COLOR,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
         flex: 1,
     },
     row: {
