@@ -1,29 +1,29 @@
 import * as React from 'react';
-import { Image, StyleSheet, StyleProp, ImageProperties, ImageStyle } from 'react-native';
+import { Image, StyleSheet, StyleProp, ImageStyle, ImageProps, ImageBackground } from 'react-native';
 
-import { ImageData } from '../models/ImageData';
+import { ImageData, BundledImage } from '../models/ImageData';
 import { ModelHelper } from '../models/ModelHelper';
+import { getImageSource } from '../helpers/imageDataHelpers';
+import { ChildrenProps } from '../ui/misc/ChildrenProps';
 
-export interface StateProps extends ImageProperties {
+export interface StateProps extends ImageProps {
     source: ImageData;
-    defaultImage?: number;
-    style: StyleProp<ImageStyle>;
+    defaultImage?: BundledImage;
+    style?: StyleProp<ImageStyle>;
     modelHelper: ModelHelper;
+    background?: boolean;
 }
 
 export interface DispatchProps {
 }
 
-export type Props = StateProps & DispatchProps;
+export type Props = StateProps & DispatchProps & Partial<ChildrenProps>;
 
 export interface State {
 }
 
 export const ImageDataView = (props: Props) => {
-    const sourceImageUri = props.modelHelper.getImageUri(props.source);
-    const source = sourceImageUri !== '' || props.defaultImage == null
-        ? { uri: sourceImageUri }
-        : props.defaultImage;
+    const source = getImageSource(props.source, props.modelHelper, props.defaultImage);
     const width = props.style
         ? StyleSheet.flatten(props.style).width != null
             ? StyleSheet.flatten(props.style).width
@@ -34,15 +34,32 @@ export const ImageDataView = (props: Props) => {
             ? StyleSheet.flatten(props.style).height
             : props.source.height
         : props.source.height;
-    return (
-        <Image
-            source={source}
-            style={[props.style, {
-                width: width,
-                height: height,
-            }]}
-        />
-    );
+    if (props.background === true) {
+        return (
+            <ImageBackground
+                {...props}
+                source={source}
+                style={[props.style, {
+                    width: width,
+                    height: height,
+                }]}
+            >
+                {props.children}
+            </ImageBackground>
+        );
+    } else {
+        return (
+            <Image
+                {...props}
+                source={source}
+                style={[props.style, {
+                    width: width,
+                    height: height,
+                }]}
+            />
+        );
+    }
+
 };
 
 const styles = StyleSheet.create({
