@@ -5,15 +5,15 @@ import { View,
     FlatList,
     Platform,
     Dimensions,
-    SafeAreaView,
 } from 'react-native';
 import Ionicon from 'react-native-vector-icons/Ionicons';
 
 import { NavigationHeader } from './NavigationHeader';
 import { clearLog, filteredLog, setLogFilter } from '../log';
-import { Colors, DefaultTabBarHeight } from '../styles';
+import { Colors, ComponentColors, DefaultTabBarHeight } from '../styles';
 import { SimpleTextInput } from './SimpleTextInput';
 import { TypedNavigation } from '../helpers/navigation';
+import { FragmentSafeAreaViewWithoutTabBar } from '../ui/misc/FragmentSafeAreaView';
 
 export interface StateProps {
     currentTimestamp: number;
@@ -39,7 +39,7 @@ export class LogViewer extends React.PureComponent<Props> {
     }
 
     public render = () => (
-        <SafeAreaView style={styles.mainContainer}>
+        <FragmentSafeAreaViewWithoutTabBar>
             <NavigationHeader
                 navigation={this.props.navigation}
                 rightButton1={{
@@ -78,23 +78,19 @@ export class LogViewer extends React.PureComponent<Props> {
                 }
                 keyExtractor={(item) => item[0]}
             />
-        </SafeAreaView>
+        </FragmentSafeAreaViewWithoutTabBar>
     )
 }
 
 const fontFamily = Platform.OS === 'ios' ? 'Courier' : 'monospace';
 const styles = StyleSheet.create({
-    mainContainer: {
-        height: Dimensions.get('window').height - DefaultTabBarHeight + 1,
-        backgroundColor: Colors.BACKGROUND_COLOR,
-    },
     logLineContainer: {
     },
     logFilterContainer: {
         height: 40,
         paddingHorizontal: 16,
         paddingVertical: 6,
-        backgroundColor: Colors.BACKGROUND_COLOR,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
     },
     logFilterTextInputContainer: {
         borderRadius: 16,
