@@ -1,12 +1,12 @@
 import * as React from 'react';
-import { View, StyleSheet, ScrollView, SafeAreaView } from 'react-native';
+import { StyleSheet, ScrollView, Vibration } from 'react-native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
 import { Settings } from '../models/Settings';
 import { Version } from '../Version';
-import { Colors } from '../styles';
+import { Colors, ComponentColors } from '../styles';
 import { NavigationHeader } from './NavigationHeader';
-import { RowItem } from '../ui/misc/RowButton';
+import { RowItem } from '../ui/buttons/RowButton';
 import { SuperGridSectionList } from 'react-native-super-grid';
 import { ReactNativeModelHelper } from '../models/ReactNativeModelHelper';
 import { GridCard, getGridCardSize, GRID_SPACING } from '../ui/misc/GridCard';
@@ -15,6 +15,8 @@ import { RecentPostFeed } from '../social/api';
 import { TabBarPlaceholder } from '../ui/misc/TabBarPlaceholder';
 import { defaultImages } from '../defaultImages';
 import { TypedNavigation } from '../helpers/navigation';
+import { FragmentSafeAreaViewForTabBar } from '../ui/misc/FragmentSafeAreaView';
+import { TouchableView } from './TouchableView';
 
 export interface StateProps {
     navigation: TypedNavigation;
@@ -30,7 +32,7 @@ export interface DispatchProps {
 
 type Props = StateProps & DispatchProps;
 
-const YOUR_FEEDS = 'YOUR FEEDS';
+const YOUR_FEEDS = 'YOUR CHANNELS';
 const PREFERENCES_LABEL = 'PREFERENCES';
 
 export const SettingsEditor = (props: Props) => {
@@ -38,13 +40,13 @@ export const SettingsEditor = (props: Props) => {
     const modelHelper = new ReactNativeModelHelper(props.settings.swarmGatewayAddress);
     const itemDimension = getGridCardSize();
     return (
-        <SafeAreaView style={{ backgroundColor: Colors.WHITE, flex: 1 }}>
+        <FragmentSafeAreaViewForTabBar>
             <NavigationHeader
                 title='Settings'
             />
-            <ScrollView style={{ backgroundColor: Colors.BACKGROUND_COLOR }}>
+            <ScrollView style={{ backgroundColor: ComponentColors.BACKGROUND_COLOR }}>
                 <SuperGridSectionList
-                    style={{ flex: 1, backgroundColor: Colors.BACKGROUND_COLOR }}
+                    style={{ flex: 1, backgroundColor: ComponentColors.BACKGROUND_COLOR }}
                     spacing={GRID_SPACING}
                     fixed={true}
                     itemDimension={itemDimension}
@@ -56,10 +58,10 @@ export const SettingsEditor = (props: Props) => {
                         return (
                                 <GridCard
                                     title={item.name}
-                                    imageUri={modelHelper.getImageUri(item.authorImage)}
+                                    image={item.authorImage}
                                     onPress={() => props.navigation.navigate('FeedSettings', { feed: item as any })}
                                     size={itemDimension}
-                                    defaultImage={defaultImages.userCircle}
+                                    defaultImage={defaultImages.defaultUser}
                                     modelHelper={modelHelper}
                                 />
                         );
@@ -82,30 +84,22 @@ export const SettingsEditor = (props: Props) => {
                     buttonStyle='switch'
                 />
                 <RowItem
-                    title='Show square images'
-                    switchState={props.settings.showSquareImages}
-                    onSwitchValueChange={props.onShowSquareImagesValueChange}
-                    buttonStyle='switch'
-                />
-                <RowItem
-                    title='Filters'
-                    buttonStyle='navigate'
-                    onPress={() => props.navigation.navigate('FilterListEditorContainer', {})}
-                />
-                <RowItem
                     title='Send bug report'
                     buttonStyle='navigate'
                     onPress={() => props.navigation.navigate('BugReportView', {})}
-                    />
-                <RowItem
-                    title={version}
-                    buttonStyle='none'
-                    onLongPress={() => props.onShowDebugMenuValueChange(!props.settings.showDebugMenu)}
                 />
+                <TouchableView
+                    onLongPress={() => {
+                        Vibration.vibrate(500, false);
+                        props.onShowDebugMenuValueChange(!props.settings.showDebugMenu);
+                    }}
+                >
+                    <RegularText style={styles.versionLabel}>{version}</RegularText>
+                </TouchableView>
                 { props.settings.showDebugMenu &&
                 <RowItem
                     icon={
-                        <Ionicons name='md-bug' size={24} color={Colors.GRAY}/>
+                        <Ionicons name='md-bug' size={24} color={ComponentColors.BUTTON_COLOR}/>
                     }
                     title='Debug menu'
                     buttonStyle='navigate'
@@ -114,7 +108,7 @@ export const SettingsEditor = (props: Props) => {
                 }
             </ScrollView>
             <TabBarPlaceholder/>
-        </SafeAreaView>
+        </FragmentSafeAreaViewForTabBar>
     );
 };
 
@@ -124,5 +118,12 @@ const styles = StyleSheet.create({
         paddingTop: 20,
         paddingBottom: 7,
         color: Colors.GRAY,
+    },
+    versionLabel: {
+        color: ComponentColors.HINT_TEXT_COLOR,
+        paddingTop: 25,
+        paddingBottom: 10,
+        paddingLeft: 10,
+        fontSize: 14,
     },
 });
