@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { AppState } from '../reducers/AppState';
-import { Actions } from '../actions/Actions';
+import { AsyncActions } from '../actions/Actions';
 import { StateProps, DispatchProps, IdentitySettings } from '../components/IdentitySettings';
 import { ImageData} from '../models/ImageData';
 import { TypedNavigation } from '../helpers/navigation';
@@ -20,10 +20,10 @@ export const mapStateToProps = (state: AppState, ownProps: { navigation: TypedNa
 const mapDispatchToProps = (dispatch: any): DispatchProps => {
     return {
         onUpdateAuthor: (text: string) => {
-            dispatch(Actions.updateAuthorName(text));
+            dispatch(AsyncActions.updateProfileName(text));
         },
         onUpdatePicture: (image: ImageData) => {
-            dispatch(Actions.updateAuthorImage(image));
+            dispatch(AsyncActions.updateProfileImage(image));
         },
     };
 };
