@@ -7,13 +7,14 @@ import { TypedNavigation } from '../helpers/navigation';
 const mapStateToProps = (state: AppState, ownProps: { navigation: TypedNavigation }): StateProps => {
     return {
         navigation: ownProps.navigation,
+        swarmGatewayAddress: state.settings.swarmGatewayAddress,
     };
 };
 
 const mapDispatchToProps = (dispatch: any): DispatchProps => {
     return {
-        onRestoreData: (data: string, secretHex: string) => {
-            dispatch(Actions.AsyncActions.restoreFromBackup(data, secretHex));
+        onRestoreData: (appState: AppState) => {
+            dispatch(Actions.AsyncActions.restoreAppStateFromBackup(appState));
         },
     };
 };
