@@ -1,11 +1,13 @@
 import { ContentFilter } from '../models/ContentFilter';
 import { Feed } from '../models/Feed';
-import { SubCategory, NewsSource } from '../models/recommendation/NewsSource';
+import { SubCategoryMap } from '../models/recommendation/NewsSource';
 import { LocalFeed } from '../social/api';
 
 export interface Routes {
     App: {};
     Loading: {};
+    Onboarding: {};
+    ProfileOnboarding: {};
     Welcome: {};
     Post: {};
     Root: {};
@@ -36,6 +38,9 @@ export interface Routes {
     FeedInfo: {
         feed: Feed;
     };
+    FeedInfoDeepLink: {
+        feedUrl: string;
+    };
     EditFilter: {
         filter: ContentFilter,
     };
@@ -51,13 +56,13 @@ export interface Routes {
     };
     SubCategoriesContainer: {
         title: string,
-        subCategories: SubCategory[],
+        subCategories: SubCategoryMap<Feed>,
     };
     NewsSourceFeed: {
         feed: Feed,
     };
     NewsSourceGridContainer: {
-        newsSources: NewsSource[],
+        feeds: Feed[],
         subCategoryName: string,
     };
     YourTab: {};
@@ -67,7 +72,7 @@ export interface TypedNavigation {
     goBack: <K extends keyof Routes>(routeKey?: K | null) => boolean;
     navigate: <K extends keyof Routes>(routeKey: K, params: Routes[K]) => boolean;
     pop: (n?: number, params?: { immediate?: boolean }) => boolean;
-    getParam: <K extends keyof Routes, P extends keyof Routes[K]>(param: P) => K[P];
+    popToTop: () => void;
+    getParam: <K extends keyof Routes, P extends keyof Routes[K]>(param: P) => Routes[K][P];
     setParams: <K extends keyof Routes>(newParams: Routes[K]) => boolean;
 }
-//         navigation.setParams<Routes['FeedInfo']>({ feed: updatedFeed });
