@@ -2,8 +2,9 @@
 import * as RNFS from 'react-native-fs';
 
 import { ModelHelper } from './ModelHelper';
-import { ImageData } from './ImageData';
+import { ImageData, BundledImage } from './ImageData';
 import { getSwarmGatewayUrl } from '../swarm/Swarm';
+import { isBundledImage } from '../helpers/imageDataHelpers';
 
 const FILE_PROTOCOL = 'file://';
 
@@ -19,7 +20,10 @@ export class ReactNativeModelHelper implements ModelHelper {
         return documentPath + localPath;
     }
 
-    public getImageUri(image: ImageData): string {
+    public getImageUri(image: ImageData): string | BundledImage {
+        if (isBundledImage(image.localPath)) {
+            return image.localPath;
+        }
         if (image.localPath != null) {
             return this.getLocalPath(image.localPath);
         }
