@@ -1,3 +1,5 @@
+import { Debug } from '../Debug';
+
 const keyBlacklist = new Set<string>()
     .add('localPath')
     .add('privateKey')
@@ -5,8 +7,18 @@ const keyBlacklist = new Set<string>()
     ;
 
 export const serialize = (data: any): string => {
-    return JSON.stringify(data, (key, value) =>
-        (key.startsWith('_') || keyBlacklist.has(key)) ? undefined : value);
+    try {
+        const serializedData = JSON.stringify(data, (key, value) => {
+            return (typeof key === 'string' && (key.startsWith('_') || keyBlacklist.has(key)))
+                ? undefined
+                : value
+            ;
+        });
+        return serializedData;
+    } catch (e) {
+        Debug.log('serialize', 'e', e, 'data', data);
+        throw e;
+    }
 };
 
 export const deserialize = (data: string): any => {
