@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { StyleSheet, View } from 'react-native';
-import { Colors } from '../../styles';
+import { Colors, ComponentColors } from '../../styles';
 import { BoldText, RegularText } from './text';
 
 interface Props {
@@ -18,7 +18,7 @@ export const PlaceholderCard = (props: Props) => {
                 {props.image}
             </View>}
             <View style={styles.textContainer}>
-                <BoldText style={styles.text}>{props.boldText}</BoldText>
+                <BoldText style={styles.boldText}>{props.boldText}</BoldText>
                 <RegularText style={styles.text}>{props.regularText}</RegularText>
             </View>
 
@@ -28,24 +28,30 @@ export const PlaceholderCard = (props: Props) => {
 
 const styles = StyleSheet.create({
     container: {
-        backgroundColor: Colors.WHITE,
-        marginBottom: 20,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
+        paddingBottom: 20,
     },
     textContainer: {
         paddingHorizontal: 20,
+    },
+    boldText: {
+        textAlign: 'center',
+        paddingTop: 10,
+        fontSize: 18,
+        color: Colors.BLACK,
     },
     text: {
         textAlign: 'center',
         paddingTop: 10,
         fontSize: 14,
-        color: Colors.DARK_GRAY,
+        color: Colors.BLACK,
     },
     imageContainer: {
         marginTop: 10,
         alignItems: 'center',
     },
     topPlaceholder: {
-        backgroundColor: Colors.BACKGROUND_COLOR,
+        backgroundColor: ComponentColors.BACKGROUND_COLOR,
         height: 10,
     },
 });
