@@ -1,21 +1,19 @@
 import * as React from 'react';
 import { FlatGrid } from 'react-native-super-grid';
-import { GridCard } from '../../misc/GridCard';
+import { GridCard, getGridCardSize } from '../../misc/GridCard';
 import { ReactNativeModelHelper } from '../../../models/ReactNativeModelHelper';
-import { View, SafeAreaView, ActivityIndicator, StyleSheet } from 'react-native';
-import { Colors } from '../../../styles';
+import { ComponentColors } from '../../../styles';
 import { NavigationHeader } from '../../../components/NavigationHeader';
-import { NewsSource } from '../../../models/recommendation/NewsSource';
-import { RSSFeedManager } from '../../../RSSPostManager';
-import { Debug } from '../../../Debug';
 import { Feed } from '../../../models/Feed';
 import { TypedNavigation } from '../../../helpers/navigation';
 import { TabBarPlaceholder } from '../../misc/TabBarPlaceholder';
+import { FragmentSafeAreaViewWithoutTabBar } from '../../misc/FragmentSafeAreaView';
+import { getFeedImage } from '../../../helpers/feedHelpers';
 
 export interface StateProps {
     gatewayAddress: string;
     subCategoryName: string;
-    newsSource: NewsSource[];
+    feeds: Feed[];
     navigation: TypedNavigation;
 }
 
@@ -23,101 +21,39 @@ export interface DispatchProps {
     downloadPostsForNewsSource: (feed: Feed) => void;
 }
 
-interface State {
-    feeds: Feed[];
-    feedLoadCounter: number;
-}
-
-export class NewsSourceGridScreen extends React.Component<StateProps & DispatchProps, State> {
-    public state: State = {
-        feeds: [],
-        feedLoadCounter: 0,
-    };
-
-    public render() {
-        const modelHelper = new ReactNativeModelHelper(this.props.gatewayAddress);
-        return (
-            <SafeAreaView style={{ backgroundColor: Colors.WHITE, flex: 1 }}>
-                <NavigationHeader title={this.props.subCategoryName} navigation={this.props.navigation}/>
-                {this.state.feeds.length > 0 &&
-                    <FlatGrid
-                       style={{ flex: 1, backgroundColor: Colors.BACKGROUND_COLOR }}
-                       spacing={10}
-                       fixed={true}
-                       itemDimension={170}
-                       items={this.state.feeds}
-                       renderItem={({ item }: any) => {
-                           const imageUri = item.authorImage ? modelHelper.getImageUri(item.authorImage) : item.favicon;
-                           return (
-                               <GridCard
-                                   title={item.name}
-                                   imageUri={imageUri}
-                                   onPress={() => {
-                                       this.props.downloadPostsForNewsSource(item);
-                                       this.props.navigation.navigate('NewsSourceFeed', {
-                                           feed: item,
-                                       });
-                                   }}
-                                   modelHelper={modelHelper}
-                                   size={170}
-                               />
-                           );
-                       }}
-                   />
-                }
-                {this.state.feedLoadCounter !== this.props.newsSource.length &&
-                    <View style={styles.activityIndicatorContainer}>
-                        <ActivityIndicator style={styles.activityIndicator} size='large'/>
-                    </View>
-                }
-                <TabBarPlaceholder color={Colors.BACKGROUND_COLOR}/>
-            </SafeAreaView>
-        );
-    }
-
-    public componentDidMount() {
-        this.props.newsSource.forEach(newsSource => {
-            fetchRSSFeedFromUrl(newsSource.url)
-                .then(feed => {
-                    if (feed != null) {
-                        this.setState((prevState: State) => {
-                            return {
-                                feeds: prevState.feeds.concat(feed),
-                                feedLoadCounter: prevState.feedLoadCounter + 1,
-                            };
-                        });
-                    } else {
-                        Debug.log(`failed to load ${newsSource.url}`);
-                        this.setState((prevState: State) => {
-                            return {
-                                feedLoadCounter: prevState.feedLoadCounter + 1,
-                            };
-                        });
-                    }
-                })
-            ;
-        });
-    }
-}
-
-const fetchRSSFeedFromUrl = async (url: string): Promise<Feed | null> => {
-    try {
-        const feed = await RSSFeedManager.fetchFeedFromUrl(url);
-        Debug.log('fetchFeedFromUrl: feed: ', feed);
-        return feed;
-    } catch (e) {
-        Debug.log(e);
-        return null;
-    }
+export const NewsSourceGridScreen = (props: StateProps & DispatchProps) => {
+    const modelHelper = new ReactNativeModelHelper(props.gatewayAddress);
+    const itemDimension = getGridCardSize();
+    return (
+        <FragmentSafeAreaViewWithoutTabBar>
+            <NavigationHeader title={props.subCategoryName} navigation={props.navigation}/>
+            {props.feeds.length > 0 &&
+                <FlatGrid
+                    style={{ flex: 1, backgroundColor: ComponentColors.BACKGROUND_COLOR }}
+                    spacing={10}
+                    fixed={true}
+                    itemDimension={itemDimension}
+                    items={props.feeds}
+                    renderItem={({ item }: any) => {
+                        const image = getFeedImage(item);
+                        return (
+                            <GridCard
+                                title={item.name}
+                                image={image}
+                                onPress={() => {
+                                    props.downloadPostsForNewsSource(item);
+                                    props.navigation.navigate('NewsSourceFeed', {
+                                        feed: item,
+                                    });
+                                }}
+                                modelHelper={modelHelper}
+                                size={itemDimension}
+                            />
+                        );
+                    }}
+                />
+            }
+            <TabBarPlaceholder color={ComponentColors.BACKGROUND_COLOR}/>
+        </FragmentSafeAreaViewWithoutTabBar>
+    );
 };
-
-const styles = StyleSheet.create({
-    activityIndicatorContainer: {
-        flex: 1,
-        backgroundColor: Colors.BACKGROUND_COLOR,
-    },
-    activityIndicator: {
-        paddingTop: 30,
-        backgroundColor: Colors.BACKGROUND_COLOR,
-    },
-});
