@@ -1,17 +1,19 @@
 import * as React from 'react';
 import { StyleSheet, ScrollView, SafeAreaView } from 'react-native';
-import { Colors } from '../../../styles';
+import { ComponentColors } from '../../../styles';
 import { RegularText } from '../../misc/text';
-import { SubCategory } from '../../../models/recommendation/NewsSource';
+import { SubCategoryMap } from '../../../models/recommendation/NewsSource';
 import { NavigationHeader } from '../../../components/NavigationHeader';
-import { RowItem } from '../../misc/RowButton';
+import { RowItem } from '../../buttons/RowButton';
 import { TypedNavigation } from '../../../helpers/navigation';
 import { TabBarPlaceholder } from '../../misc/TabBarPlaceholder';
+import { FragmentSafeAreaViewWithoutTabBar} from '../../misc/FragmentSafeAreaView';
+import { Feed } from '../../../models/Feed';
 
 const SUBCATEGORIES_LABEL = 'SUBCATEGORIES';
 
 export interface StateProps {
-    subCategories: SubCategory[];
+    subCategories: SubCategoryMap<Feed>;
     navigation: TypedNavigation;
     title: string;
 }
@@ -23,30 +25,32 @@ export interface OwnProps {
 export interface DispatchProps { }
 
 export const SubCategoriesScreen = (props: StateProps & DispatchProps) => {
-    const subCategories = props.subCategories.map((subCategory: SubCategory) => {
+    const subCategories = Object.keys(props.subCategories).map((subCategoryName) => {
         return (
             <RowItem
-                key={subCategory.name}
-                title={subCategory.name}
+                key={subCategoryName}
+                title={subCategoryName}
                 buttonStyle='navigate'
                 onPress={() => props.navigation.navigate('NewsSourceGridContainer', {
-                    newsSources: subCategory.list,
-                    subCategoryName: subCategory.name,
+                    feeds: props.subCategories[subCategoryName],
+                    subCategoryName,
                 })}
             />
         );
     });
     return (
-        <SafeAreaView style={{ backgroundColor: Colors.WHITE, flex: 1 }}>
-            <NavigationHeader title={props.title} navigation={props.navigation}/>
-            <ScrollView style={{ backgroundColor: Colors.BACKGROUND_COLOR }}>
-                <RegularText style={styles.label}>
-                    {SUBCATEGORIES_LABEL}
-                </RegularText>
-                {subCategories}
-            </ScrollView>
-            <TabBarPlaceholder color={Colors.BACKGROUND_COLOR}/>
-        </SafeAreaView>
+        <FragmentSafeAreaViewWithoutTabBar>
+            <SafeAreaView style={{flex: 1}}>
+                <NavigationHeader title={props.title} navigation={props.navigation}/>
+                <ScrollView style={{ backgroundColor: ComponentColors.BACKGROUND_COLOR }}>
+                    <RegularText style={styles.label}>
+                        {SUBCATEGORIES_LABEL}
+                    </RegularText>
+                    {subCategories}
+                </ScrollView>
+                <TabBarPlaceholder color={ComponentColors.BACKGROUND_COLOR}/>
+            </SafeAreaView>
+        </FragmentSafeAreaViewWithoutTabBar>
     );
 };
 
@@ -55,6 +59,6 @@ const styles = StyleSheet.create({
         paddingHorizontal: 10,
         paddingTop: 20,
         paddingBottom: 7,
-        color: Colors.GRAY,
+        color: ComponentColors.TEXT_COLOR,
     },
 });
